[Impr./Feat.] Improve Number API and add several number expressions. (#75)
* [Feature] Add PropertyConditionals as utility class

* [Impr./Feat.] Improve number API

* [Impr.] Improve ExprBinaryMathFunctions

* [Improvements] Improve other syntax as well

* [Improvement] Optimize imports

* [Feature] Add CondExprIsDivisible and fix bug in EffDoIf

* [Feature] Add CondExprIsPrime

* [Feature] Add CondExprIsPrime and optimisations

* [Feature] Add ExprPrime

* [Feature] Add ExprPrime

* [Fix] Fix new time implementation

* [Fix] Support variables in ExprDifference

* [Fix] Fix NPE with date parsing

* [Improve] Improve primality checks

* [Fix] Remove unnecessary code

* [Improvement] Improve primality check

* request: add requests by Syst3ms

+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprAmount.java
@@ -10,6 +10,7 @@ import io.github.syst3ms.skriptparser.parsing.ParseContext;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.Optional;
 
@@ -57,9 +58,11 @@ public class ExprAmount extends PropertyExpression<Number, Object> {
 		if (recursive) {
 			Optional<Object> var = ((Variable<?>) getOwner()).getRaw(ctx);
 			if (var.isPresent())
-				return new Number[] {getRecursiveSize((Map<String, ?>) var.get())}; // Should already be a Map
+				return new Number[] {
+						BigInteger.valueOf(getRecursiveSize((Map<String, ?>) var.get()))
+			};
 		}
-		return new Number[] {getOwner().getValues(ctx).length};
+		return new Number[] {BigInteger.valueOf(getOwner().getValues(ctx).length)};
 	}
 
 	@Override
