+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprDateTimestamp.java
@@ -8,6 +8,7 @@ import io.github.syst3ms.skriptparser.parsing.ParseContext;
 import io.github.syst3ms.skriptparser.util.SkriptDate;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
 import java.util.Optional;
 import java.util.function.Function;
 
@@ -43,8 +44,8 @@ public class ExprDateTimestamp extends PropertyExpression<Number, SkriptDate> {
 		return Optional.of(
 				dates -> new Number[] {
 					unix
-						? Math.floorDiv(dates[0].getTimestamp(), 1000)
-						: dates[0].getTimestamp()
+						? BigInteger.valueOf(Math.floorDiv(dates[0].getTimestamp(), 1000))
+						: BigInteger.valueOf(dates[0].getTimestamp())
 				}
 		);
 	}
